Use document.unload to disconnect the clients
This notifies the server of clients leaving the 
game if the client closes the tab, restarts the
browser or if it goes back to the previous site.

This however still fails to notify the server if
the clients javascript crashes by dereferencing 
undefined variables, etc.

Noticed strange behaviour when opening multiple
tabs in the same computer, sometimes the clients
crash, sometimes they don’t. Errors include 
‘cannot read property of 5 of undefined’ in 
client_game.js:159. In coordinate to cell method,
this.grid[x_co] is undefined.